pipeline oii

+++ preditor.h
@@ -6,7 +6,7 @@
 #define BHT_SIZE 32
 
 typedef struct {
-    int contador; // (0, 1, 2, 3)
+    uint8_t contador; // (0, 1, 2, 3) vulgo 2 bit
     uint16_t alvo;
 } EntradaBHT;
